Naprawa importu statków
Import + przycisk play
