+++ src/App.tsx
@@ -1,6 +1,6 @@
 import './App.css'
 import { Helmet } from 'react-helmet';
-import Statki from './Statki.tsx';
+import Statki from './statki/statki.tsx';
 import Saper from './saper/saper.tsx';
 import Snake from './snake/snake.tsx';
 import Tic_Tac_Toe from './Tic_Tac_Toe.tsx';
