+++ src/statki/statki.js
@@ -1,312 +1,316 @@
-const gamesboardContainer = document.querySelector('#gamesboard-container')
-const optionContainer = document.querySelector('.option-container')
-const flipButton = document.querySelector('#flip-button')
-const startButton = document.querySelector('#start-button')
-const infoDisplay = document.querySelector('#info')
-const turnDisplay = document.querySelector('#turn-display')
+const Statkii = () =>{
+    const gamesboardContainer = document.querySelector('#gamesboard-container')
+    const optionContainer = document.querySelector('.option-container')
+    const flipButton = document.querySelector('#flip-button')
+    const startButton = document.querySelector('#start-button')
+    const infoDisplay = document.querySelector('#info')
+    const turnDisplay = document.querySelector('#turn-display')
 
-// Option choosing
-let angle = 0
-function flip() {
-    const optionShips = Array.from(optionContainer.children)
-    angle = angle === 0 ? 90 : 0
-    optionShips.forEach(optionShip => optionShip.style.transform = `rotate(${angle}deg)`)
-}
-
-flipButton.addEventListener('click', flip);
-
-
-// Creating Boards
-const width = 10
-
-function createBoard(color, user) {
-    const gameboardContainer = document.createElement('div')
-    gameboardContainer.classList.add('game-board')
-    gameboardContainer.style.backgroundColor = color
-    gameboardContainer.id = user
-
-    for(let i = 0;i < width * width; i++)
-    {
-        const block = document.createElement('div')
-        block.classList.add('block')
-        block.id = i
-        gameboardContainer.append(block)
+    // Option choosing
+    let angle = 0
+    function flip() {
+        const optionShips = Array.from(optionContainer.children)
+        angle = angle === 0 ? 90 : 0
+        optionShips.forEach(optionShip => optionShip.style.transform = `rotate(${angle}deg)`)
     }
 
-    gamesboardContainer.append(gameboardContainer)
-}
-createBoard('#736679', 'player')
-createBoard('#301934', 'computer')
+    flipButton.addEventListener('click', flip);
 
-//Creating Ships
-class Ship
-{
-    constructor(name, length)
-    {
-        this.name = name
-        this.length = length
+
+    // Creating Boards
+    const width = 10
+
+    function createBoard(color, user) {
+        const gameboardContainer = document.createElement('div')
+        gameboardContainer.classList.add('game-board')
+        gameboardContainer.style.backgroundColor = color
+        gameboardContainer.id = user
+
+        for(let i = 0;i < width * width; i++)
+        {
+            const block = document.createElement('div')
+            block.classList.add('block')
+            block.id = i
+            gameboardContainer.append(block)
+        }
+
+        gamesboardContainer.append(gameboardContainer)
     }
-}
+    createBoard('#736679', 'player')
+    createBoard('#301934', 'computer')
 
-const destroyer = new Ship('destroyer', 2)
-const submarine = new Ship('submarine', 3)
-const cruiser = new Ship('cruiser', 3)
-const battleship = new Ship('battleship', 4)
-const carrier = new Ship('carrier', 5)
-
-const ships = [destroyer, submarine, cruiser, battleship, carrier]
-let notDropped
-
-function getValidity(allBoardBlocks, isHorizontal, startIndex, ship)
-{
-    let validStart = isHorizontal ? startIndex <= width * width - ship.length ? startIndex :
-        width * width - ship.length : 
-        //handle vertical
-        startIndex <= width * width - width * ship.length ? startIndex : 
-            startIndex - ship.length * width + width
-
-    let shipBlocks = [];
-
-    for (let i = 0; i < ship.length; i++) {
-        if (isHorizontal) {
-            shipBlocks.push(allBoardBlocks[Number(validStart) + i]);
-        } else {
-            shipBlocks.push(allBoardBlocks[Number(validStart) + i * width]);
+    //Creating Ships
+    class Ship
+    {
+        constructor(name, length)
+        {
+            this.name = name
+            this.length = length
         }
     }
 
-    let valid
+    const destroyer = new Ship('destroyer', 2)
+    const submarine = new Ship('submarine', 3)
+    const cruiser = new Ship('cruiser', 3)
+    const battleship = new Ship('battleship', 4)
+    const carrier = new Ship('carrier', 5)
 
-    if(isHorizontal) 
+    const ships = [destroyer, submarine, cruiser, battleship, carrier]
+    let notDropped
+
+    function getValidity(allBoardBlocks, isHorizontal, startIndex, ship)
     {
-        shipBlocks.every((_shipBlock, index) =>
-            valid = shipBlocks[0].id % width !== width - (shipBlocks.length - (index + 1)))
-    }
-    else
-    {
-        shipBlocks.every((_shipBlock, index) =>
-            valid = shipBlocks[0].id < 90 + (width * index + 1))
-    }
+        let validStart = isHorizontal ? startIndex <= width * width - ship.length ? startIndex :
+            width * width - ship.length : 
+            //handle vertical
+            startIndex <= width * width - width * ship.length ? startIndex : 
+                startIndex - ship.length * width + width
+
+        let shipBlocks = [];
+
+        for (let i = 0; i < ship.length; i++) {
+            if (isHorizontal) {
+                shipBlocks.push(allBoardBlocks[Number(validStart) + i]);
+            } else {
+                shipBlocks.push(allBoardBlocks[Number(validStart) + i * width]);
+            }
+        }
+
+        let valid
+
+        if(isHorizontal) 
+        {
+            shipBlocks.every((_shipBlock, index) =>
+                valid = shipBlocks[0].id % width !== width - (shipBlocks.length - (index + 1)))
+        }
+        else
+        {
+            shipBlocks.every((_shipBlock, index) =>
+                valid = shipBlocks[0].id < 90 + (width * index + 1))
+        }
 
 
-    const notTaken = shipBlocks.every(shipBlock => !shipBlock.classList.contains('taken'))
+        const notTaken = shipBlocks.every(shipBlock => !shipBlock.classList.contains('taken'))
 
-    return { shipBlocks, valid, notTaken}
-}
-
-function addShipPiece(user, ship, startId) {
-    const allBoardBlocks = document.querySelectorAll(`#${user} div`); 
-    let randomBoolean = Math.random() < 0.5;
-    let isHorizontal = user === 'player' ? angle === 0 : randomBoolean;
-    let randomStartIndex = Math.floor(Math.random() * width * width);
-
-    let startIndex = startId ? startId : randomStartIndex
-
-
-    const {shipBlocks, valid, notTaken} = getValidity(allBoardBlocks, isHorizontal, startIndex, ship)
-
-    if (valid && notTaken)
-    {
-        shipBlocks.forEach(shipBlock => {
-            shipBlock.classList.add(ship.name);
-            shipBlock.classList.add('taken');
-        })
-    }
-    else{
-        if (user === 'computer') addShipPiece(user, ship, startId)
-        if (user === 'player') notDropped = false
-    }
-}
-
-ships.forEach(ship => addShipPiece('computer', ship))
-
-
-//Drag player ships
-let draggedShip
-const optionShips = Array.from(optionContainer.children)
-optionShips.forEach(optionShips => optionShips.addEventListener('dragstart', dragStart))
-
-const allPlayerBlocks = document.querySelectorAll('#player div')
-allPlayerBlocks.forEach(playerBlock => 
-    {
-        playerBlock.addEventListener('dragover', dragOver)
-        playerBlock.addEventListener('drop', dropShip)
-    })
-
-function dragStart(e)
-{
-    notDropped = false
-    draggedShip = e.target
-}
-
-function dragOver(e)
-{
-    e.preventDefault()
-    const ship = ships[draggedShip.id]
-    highlightArea(e.target.id, ship)
-}
-
-function dropShip(e)
-{
-    const startId = e.target.id
-    const ship = ships[draggedShip.id]
-    addShipPiece('player', ship, startId)
-    if (!notDropped)
-    {
-        draggedShip.remove()
-    }
-}
-
-//Add highlight
-function highlightArea( startIndex, ship)
-{
-    const allBoardBlocks = document.querySelectorAll('#player div')
-    let isHorizontal = angle === 0
-
-    const {shipBlocks, valid, notTaken} = getValidity(allBoardBlocks, isHorizontal, startIndex, ship)
-
-    if (valid && notTaken)
-    {
-        shipBlocks.forEach(shipBlock => {
-            shipBlock.classList.add('hover')
-            setTimeout(() => shipBlock.classList.remove('hover'), 500)
-        })
+        return { shipBlocks, valid, notTaken}
     }
 
-}
+    function addShipPiece(user, ship, startId) {
+        const allBoardBlocks = document.querySelectorAll(`#${user} div`); 
+        let randomBoolean = Math.random() < 0.5;
+        let isHorizontal = user === 'player' ? angle === 0 : randomBoolean;
+        let randomStartIndex = Math.floor(Math.random() * width * width);
 
-let gameOver = false
-let playerTurn
+        let startIndex = startId ? startId : randomStartIndex
 
-//Start Game
-function startGame()
-{
-    if (playerTurn === undefined)
-    {
-        if(optionContainer.children.length != 0){
-            infoDisplay.textContent = 'Proszę najpierw wystawić wszystkie swoje statki'
+
+        const {shipBlocks, valid, notTaken} = getValidity(allBoardBlocks, isHorizontal, startIndex, ship)
+
+        if (valid && notTaken)
+        {
+            shipBlocks.forEach(shipBlock => {
+                shipBlock.classList.add(ship.name);
+                shipBlock.classList.add('taken');
+            })
         }
         else{
-            const allBoardBlocks = document.querySelectorAll('#computer div')
-            allBoardBlocks.forEach(block => block.addEventListener('click', handleClick))
-        playerTurn = true
-        turnDisplay.textContent = 'Zaczynasz!'
-        infoDisplay.textContent = 'Gra START'
+            if (user === 'computer') addShipPiece(user, ship, startId)
+            if (user === 'player') notDropped = false
         }
     }
-}
 
-startButton.addEventListener('click', startGame)
+    ships.forEach(ship => addShipPiece('computer', ship))
 
-let playerHits = []
-let computerHits = []
-const playerSunkShips = []
-const computerSunkShips = []
 
-function handleClick(e){
-    if(!gameOver){
-        if(e.target.classList.contains('taken')){
-            e.target.classList.add('boom')
-            infoDisplay.textContent = "Statek został trafiony!"
-            let classes = Array.from(e.target.classList)
-            classes = classes.filter(className => className !== 'block')
-            classes = classes.filter(className => className !== 'boom')
-            classes = classes.filter(className => className !== 'taken')
-            playerHits.push(...classes)
-            checkScore('player', playerHits, playerSunkShips)
-        }
-        if (!e.target.classList.contains('taken'))
+    //Drag player ships
+    let draggedShip
+    const optionShips = Array.from(optionContainer.children)
+    optionShips.forEach(optionShips => optionShips.addEventListener('dragstart', dragStart))
+
+    const allPlayerBlocks = document.querySelectorAll('#player div')
+    allPlayerBlocks.forEach(playerBlock => 
         {
-            infoDisplay.textContent = "Pudło!"
-            e.target.classList.add('empty')
+            playerBlock.addEventListener('dragover', dragOver)
+            playerBlock.addEventListener('drop', dropShip)
+        })
+
+    function dragStart(e)
+    {
+        notDropped = false
+        draggedShip = e.target
+    }
+
+    function dragOver(e)
+    {
+        e.preventDefault()
+        const ship = ships[draggedShip.id]
+        highlightArea(e.target.id, ship)
+    }
+
+    function dropShip(e)
+    {
+        const startId = e.target.id
+        const ship = ships[draggedShip.id]
+        addShipPiece('player', ship, startId)
+        if (!notDropped)
+        {
+            draggedShip.remove()
         }
-        playerTurn = false
-        const allBoardBlocks = document.querySelectorAll('#computer div')
-        allBoardBlocks.forEach(block => block.replaceWith(block.cloneNode(true)))
-        setTimeout(computerGo, 3000)
+    }
+
+    //Add highlight
+    function highlightArea( startIndex, ship)
+    {
+        const allBoardBlocks = document.querySelectorAll('#player div')
+        let isHorizontal = angle === 0
+
+        const {shipBlocks, valid, notTaken} = getValidity(allBoardBlocks, isHorizontal, startIndex, ship)
+
+        if (valid && notTaken)
+        {
+            shipBlocks.forEach(shipBlock => {
+                shipBlock.classList.add('hover')
+                setTimeout(() => shipBlock.classList.remove('hover'), 500)
+            })
+        }
 
     }
-}
 
-//Define the computer turn
-function computerGo(){
-    if(!gameOver)
+    let gameOver = false
+    let playerTurn
+
+    //Start Game
+    function startGame()
     {
-        turnDisplay.textContent = 'Tura Komputera'
-        infoDisplay.textContent = 'Komputer myśli nad ruchem...'
-
-        setTimeout(() => {
-            let randomGo = Math.floor(Math.random() * width * width)
-            const allBoardBlocks = document.querySelectorAll('#player div')
-
-            if(allBoardBlocks[randomGo].classList.contains('taken') &&
-            allBoardBlocks[randomGo].classList.contains('boom')
-            ){
-                computerGo()
-                return
+        if (playerTurn === undefined)
+        {
+            if(optionContainer.children.length != 0){
+                infoDisplay.textContent = 'Proszę najpierw wystawić wszystkie swoje statki'
             }
-            else if(allBoardBlocks[randomGo].classList.contains('taken') &&
-            !allBoardBlocks[randomGo].classList.contains('boom')
-            ){
-                allBoardBlocks[randomGo].classList.add('boom')
-                infoDisplay.textContent = 'Komputer trafił twój statek!'
-                let classes = Array.from(allBoardBlocks[randomGo].classList)
+            else{
+                const allBoardBlocks = document.querySelectorAll('#computer div')
+                allBoardBlocks.forEach(block => block.addEventListener('click', handleClick))
+            playerTurn = true
+            turnDisplay.textContent = 'Zaczynasz!'
+            infoDisplay.textContent = 'Gra START'
+            }
+        }
+    }
+
+    startButton.addEventListener('click', startGame)
+
+    let playerHits = []
+    let computerHits = []
+    const playerSunkShips = []
+    const computerSunkShips = []
+
+    function handleClick(e){
+        if(!gameOver){
+            if(e.target.classList.contains('taken')){
+                e.target.classList.add('boom')
+                infoDisplay.textContent = "Statek został trafiony!"
+                let classes = Array.from(e.target.classList)
                 classes = classes.filter(className => className !== 'block')
                 classes = classes.filter(className => className !== 'boom')
                 classes = classes.filter(className => className !== 'taken')
-                computerHits.push(...classes)
-                checkScore('computer', computerHits, computerSunkShips)
+                playerHits.push(...classes)
+                checkScore('player', playerHits, playerSunkShips)
             }
-            else{
-                infoDisplay.textContent = 'Pudło ze strony komputera'
-                allBoardBlocks[randomGo].classList.add('empty')
+            if (!e.target.classList.contains('taken'))
+            {
+                infoDisplay.textContent = "Pudło!"
+                e.target.classList.add('empty')
             }
-        }, 3000)
-
-        setTimeout(() => {
-            playerTurn = true
-            turnDisplay.textContent = 'Twoja tura!'
-            infoDisplay.textContent = 'Spróbuj trafić przeciwnika'
+            playerTurn = false
             const allBoardBlocks = document.querySelectorAll('#computer div')
-            allBoardBlocks.forEach(block => block.addEventListener('click', handleClick))
-        }, 6000);
+            allBoardBlocks.forEach(block => block.replaceWith(block.cloneNode(true)))
+            setTimeout(computerGo, 3000)
 
-    }
-}
-
-function checkScore(user, userHits, userSunkShips)
-{
-    function checkShip(shipName, shipLength)
-    {
-        if(
-            userHits.filter(storedShipName => storedShipName === shipName).length ===shipLength
-        ){
-            if(user === 'player'){
-                infoDisplay.textContent = `Zatopiłeś ${shipName} Komputera!`
-                playerHits = userHits.filter(storedShipName => storedShipName !== shipName)
-            }
-            if(user === 'computer'){
-                infoDisplay.textContent = `Zatopił ${shipName} Gracza`
-                computerHits = userHits.filter(storedShipName => storedShipName !== shipName)
-            }
-            userSunkShips.push(shipName)
         }
     }
 
-    checkShip('destroyer', 2)
-    checkShip('submarine', 3)
-    checkShip('cruiser', 3)
-    checkShip('battleship', 4)
-    checkShip('carrier', 5)
+    //Define the computer turn
+    function computerGo(){
+        if(!gameOver)
+        {
+            turnDisplay.textContent = 'Tura Komputera'
+            infoDisplay.textContent = 'Komputer myśli nad ruchem...'
 
-    console.log('playerHits', playerHits)
-    console.log('playerSunkShips', playerSunkShips)
+            setTimeout(() => {
+                let randomGo = Math.floor(Math.random() * width * width)
+                const allBoardBlocks = document.querySelectorAll('#player div')
 
-    if(playerSunkShips.length === 5) {
-        infoDisplay.textContent = 'Zatopiłeś wszystkie statki wroga. WYGRAŁEŚ!!!!'
+                if(allBoardBlocks[randomGo].classList.contains('taken') &&
+                allBoardBlocks[randomGo].classList.contains('boom')
+                ){
+                    computerGo()
+                    return
+                }
+                else if(allBoardBlocks[randomGo].classList.contains('taken') &&
+                !allBoardBlocks[randomGo].classList.contains('boom')
+                ){
+                    allBoardBlocks[randomGo].classList.add('boom')
+                    infoDisplay.textContent = 'Komputer trafił twój statek!'
+                    let classes = Array.from(allBoardBlocks[randomGo].classList)
+                    classes = classes.filter(className => className !== 'block')
+                    classes = classes.filter(className => className !== 'boom')
+                    classes = classes.filter(className => className !== 'taken')
+                    computerHits.push(...classes)
+                    checkScore('computer', computerHits, computerSunkShips)
+                }
+                else{
+                    infoDisplay.textContent = 'Pudło ze strony komputera'
+                    allBoardBlocks[randomGo].classList.add('empty')
+                }
+            }, 3000)
+
+            setTimeout(() => {
+                playerTurn = true
+                turnDisplay.textContent = 'Twoja tura!'
+                infoDisplay.textContent = 'Spróbuj trafić przeciwnika'
+                const allBoardBlocks = document.querySelectorAll('#computer div')
+                allBoardBlocks.forEach(block => block.addEventListener('click', handleClick))
+            }, 6000);
+
+        }
     }
-    if(computerSunkShips.length === 5) {
-        infoDisplay.textContent = 'Komputer zatopił wszystkie twoje statki. PORAŻKA!!!!'
-        gameOver = true
+
+    function checkScore(user, userHits, userSunkShips)
+    {
+        function checkShip(shipName, shipLength)
+        {
+            if(
+                userHits.filter(storedShipName => storedShipName === shipName).length ===shipLength
+            ){
+                if(user === 'player'){
+                    infoDisplay.textContent = `Zatopiłeś ${shipName} Komputera!`
+                    playerHits = userHits.filter(storedShipName => storedShipName !== shipName)
+                }
+                if(user === 'computer'){
+                    infoDisplay.textContent = `Zatopił ${shipName} Gracza`
+                    computerHits = userHits.filter(storedShipName => storedShipName !== shipName)
+                }
+                userSunkShips.push(shipName)
+            }
+        }
+
+        checkShip('destroyer', 2)
+        checkShip('submarine', 3)
+        checkShip('cruiser', 3)
+        checkShip('battleship', 4)
+        checkShip('carrier', 5)
+
+        console.log('playerHits', playerHits)
+        console.log('playerSunkShips', playerSunkShips)
+
+        if(playerSunkShips.length === 5) {
+            infoDisplay.textContent = 'Zatopiłeś wszystkie statki wroga. WYGRAŁEŚ!!!!'
+        }
+        if(computerSunkShips.length === 5) {
+            infoDisplay.textContent = 'Komputer zatopił wszystkie twoje statki. PORAŻKA!!!!'
+            gameOver = true
+        }
     }
 }
+
+export default Statkii;
